ch(implement-feedback): implement feedback across project
- implement all feedback on UI
- refactor sql arrow functions to be a class

[Finishes #160189517]
