+++ UI/login.html
@@ -35,7 +35,7 @@
     </div>
 </div>
 
-    <div class = "container image-background" style = "margin: 0 auto; height: 800px;">
+    <div class = "container image-background ask-bg" style = "margin: 0 auto;">
     
         <div class = "row ">
