+++ UI/profile.html
@@ -137,8 +137,8 @@
                   </div>
                   <div class = "row mt-2 pd-1">
                     <div class = "col"> <span> 10 Answers</span></div>
-                    <div class = "col"><span> <i class="fas fa-arrow-alt-circle-up"></i></span>&nbsp 200<span></span></div>
-                    <div class = "col"><span> <i class="fas fa-arrow-alt-circle-down "></i></span>&nbsp 200<span><span></span></div>
+                    <div class = "col"><span> <i class="fas fa-thumbs-up"></i></span>&nbsp 200<span></span></div>
+                    <div class = "col"><span> <i class="fas fa-thumbs-down "></i></span>&nbsp 200<span><span></span></div>
                     <div class = "col"><span></span><span></span><button type="answer">Answer</button></div>
                   </div>
                 </div>
@@ -163,8 +163,8 @@
                       </div>
                       <div class = "row mt-2 pd-1">
                         <div class = "col"> <span> 10 Answers</span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-up"></i></span>&nbsp 200<span></span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-down "></i></span>&nbsp 200<span><span></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-up"></i></span>&nbsp 200<span></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-down "></i></span>&nbsp 200<span><span></span></div>
                         <div class = "col"><span></span><span></span><button type="answer">Answer</button></div>
                       </div>
                     </div>
@@ -189,8 +189,8 @@
                       </div>
                       <div class = "row mt-2 pd-1">
                         <div class = "col"> <span> 10 Answers</span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-up"></i></span>&nbsp 200<span></span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-down "></i></span>&nbsp 200<span><span></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-up"></i></span>&nbsp 200<span></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-down "></i></span>&nbsp 200<span><span></span></div>
                         <div class = "col"><span></span><span></span><button type="answer">Answer</button></div>
                       </div>
                     </div>
@@ -226,8 +226,8 @@
                   </div>
                   <div class = "row mt-2 pd-1">
                     <div class = "col"> <span> 10 Answers</span></div>
-                    <div class = "col"><span> <i class="fas fa-arrow-alt-circle-up"></i></span>&nbsp 200<span></span></div>
-                    <div class = "col"><span> <i class="fas fa-arrow-alt-circle-down "></i></span>&nbsp 200<span><span></span></div>
+                    <div class = "col"><span> <i class="fas fa-thumbs-up"></i></span>&nbsp 200<span></span></div>
+                    <div class = "col"><span> <i class="fas fa-thumbs-down "></i></span>&nbsp 200<span><span></span></div>
                     <div class = "col"><span></span><span></span><button type="answer">Answer</button></div>
                   </div>
                 </div>
@@ -252,8 +252,8 @@
                       </div>
                       <div class = "row mt-2 pd-1">
                         <div class = "col"> <span> 10 Answers</span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-up"></i></span>&nbsp 200<span></span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-down "></i></span>&nbsp 200<span><span></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-up"></i></span>&nbsp 200<span></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-down "></i></span>&nbsp 200<span><span></span></div>
                         <div class = "col"><span></span><span></span><button type="answer">Answer</button></div>
                       </div>
                     </div>
@@ -278,8 +278,8 @@
                       </div>
                       <div class = "row mt-2 pd-1">
                         <div class = "col"> <span> 10 Answers</span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-up"></i></span>&nbsp 200<span></span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-down "></i></span>&nbsp 200<span><span></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-up"></i></span>&nbsp 200<span></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-down "></i></span>&nbsp 200<span><span></span></div>
                         <div class = "col"><span></span><span></span><button type="answer">Answer</button></div>
                       </div>
                     </div>
@@ -290,7 +290,7 @@
 
     </div>
 
-<div class = "container " style = "text-align: center;color:white; background-color:#4d4d4d; padding: 1.5%;" >
+<div class = "container  footer"  >
    &copy 2018 StackOverFlow-Lite
 
 </div>
