+++ UI/questions.html
@@ -45,7 +45,7 @@
                 <div class = "underline">&nbsp</div>
                 <div class = "row">
                   <div class = "col-5">My android device is not responding since I installed its latest operating system
-                    <div class = "mt-4">Asked by Augustine Ezinwa  &nbsp <span class = "darkgray" > 15 mins ago</span></div>
+                    <div class = "mt-4 ft">Asked by Augustine Ezinwa  &nbsp <span class = "darkgray" > 15 mins ago</span></div>
                   </div>
                   <div class = "col-2">
                     <div class = "row">
@@ -64,7 +64,7 @@
                 <div class = "row">
                   <div class = "col-5">Consider resetting your device to factory settings. 
                     <div class = "row mt-4">
-                      <div class = "col"><div class = "">Answered by Augustine Ezinwa  &nbsp <span class = "darkgray" >15 mins ago</span></div></div>
+                      <div class = "col"><div class = "ft">Answered by Augustine Ezinwa  &nbsp <span class = "darkgray" >15 mins ago</span></div></div>
                       <div class = "col ">
                         <div class = "row">
                           <div class = "col"> <button type =comment>Comment</button></div>
@@ -85,9 +85,9 @@
                     </div>
 
                     <div class = "row">
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-up blue resize"></i></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-up blue resize"></i></span></div>
                         <div class = "col"><span> <i class="fas fa-star blue resize"></i></span></div>
-                        <div class = "col"><span> <i class="fas fa-arrow-alt-circle-down blue resize"></i></span></div>
+                        <div class = "col"><span> <i class="fas fa-thumbs-down blue resize"></i></span></div>
                     </div>
                   </div>
 
@@ -103,7 +103,7 @@
                 <div class = "row">
                     <div class = "col-5">Consider resetting your device to factory settings. 
                       <div class = "row mt-4">
-                        <div class = "col"><div class = "">Answered by Augustine Ezinwa  &nbsp <span class = "darkgray">15 mins ago</span></div></div>
+                        <div class = "col"><div class = "ft">Answered by Augustine Ezinwa  &nbsp <span class = "darkgray">15 mins ago</span></div></div>
                         <div class = "col ">
                           <div class = "row">
                             <div class = "col"> <button type ="comment">Comment</button></div>
@@ -124,9 +124,9 @@
                       </div>
   
                       <div class = "row">
-                          <div class = "col"><span> <i class="fas fa-arrow-alt-circle-up blue resize"></i></span></div>
+                          <div class = "col"><span> <i class="fas fa-thumbs-up blue resize"></i></span></div>
                           <div class = "col"><span> <i class="fas fa-star blue resize"></i></span></div>
-                          <div class = "col"><span> <i class="fas fa-arrow-alt-circle-down blue resize"></i></span></div>
+                          <div class = "col"><span> <i class="fas fa-thumbs-down blue resize"></i></span></div>
                       </div>
                     </div>
                   </div>
