+++ UI/styles/index.css
@@ -10,7 +10,7 @@ body {
   height: 400px;
   width: 100%;
   margin:0 auto;
- 
+  box-shadow:inset 0 0 0 2000px rgba(0,0,0,0.3);
   
 }
 
@@ -263,7 +263,10 @@ input[type=text], input[type=password] {
 
 .ask-bg {
   margin: 0 auto;
-   height: 800px;
+   height:1000px;
+}
+.ft {
+  font-size: 14.48px;
 }
 .txtarea {
   height: 200px;
@@ -288,6 +291,7 @@ button[type=comment] {
    color: hotpink;
     border: 1px solid hotpink;
 }
+
 
 @media screen and (max-width: 768px) {
   .col,.col-2, .col-5, .col-6, .col-3 {
@@ -362,7 +366,8 @@ color: darkgray;
   background-color: grey;
 }
 .profile-height {
-  height: 180vh;
+  height: 110%;
+
 }
 
 .adjust {
@@ -408,6 +413,8 @@ color: darkgray;
   margin-top: 10%;
   
 }
-
+.ask-bg {
+  height: 800px;
+}
 
 }
