+++ UI/styles/nav.css
@@ -1,10 +1,11 @@
 * {
     box-sizing: border-box;
   }
-  
+  html {
+    font-family:'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+  }
   body {
     margin: 0px;
-    font-family: 'segoe ui';
   }
   .bg {
       background-color: black;
@@ -14,7 +15,8 @@
     height: 53px;
     width: 100%;
     background-color: #4d4d4d;
-    position: relative;
+    position: fixed;
+    top:0;
   }
   
   .nav > .nav-header {
@@ -124,7 +126,7 @@
       height: 0px;
     }
     .nav > #nav-check:checked + .nav-links {
-      height: calc(100vh - 50px);
+      height: calc(80vh - 50px);
       overflow-y: auto;
     }
     .search-bar {
