+++ server/tests/questions/questionControllerTests.js
@@ -2,7 +2,9 @@ import chai from 'chai';
 import chaiHttp from 'chai-http';
 import app from '../../app';
 import dbConnect from '../../connections/dbConnect';
-import { getAUserQuestion } from '../../helper/sqlHelper';
+import SqlHelper from '../../helper/SqlHelper';
+
+const { getAUserQuestion } = SqlHelper;
 
 chai.use(chaiHttp);
 const should = chai.should();
